Point /press to /index.html in docker nginx config
Due to how we're handling routing in the app, /press isn't a separate
page, but rather a state of the main page. However, nginx doesn't know
that, so it's trying to resolve a file at /press that does not exist.

As with the business primer, try_files is used here to allow for
replacing the page with an actual entity at /press if needed.