Move alert banner inside header component
This makes css easier; now we can just make the entire
header component fixed instead of the nav and alert set
as fixed separately with explicit top positions. This allows
the banner to have a variable content length (multiple lines)
without having to adjust the top positioning of the nav
based on the height of the alert.