Fix flickering font weight for scroll aware links
Flickering for the scroll aware links is caused by applying the heavy font weight on
the active state, which occurs on mouse down only (it is removed on
mouse up), and the "active" class being applied when the scroll position
is finally on the section associated with the scroll aware link. Between those two
moments the active style isn't actually applicable (during scroll time). To fix this,
we skip applying the heavy font weight on the mouse down (click) for the scroll link
and save it for only when the "active" class is applied.