Add featured application section with tBTC
This section is immediately after the hero heading and has links to both the
tBTC site and the introductory Keep blog post for tBTC. It is loosely
structured to be able to support multiple applications in a similar structure.

Currently the tBTC app image is managed as a background image rather than an
img element, mostly due to difficulty positioning the img element correctly, as
well as the current inability to use an SVG for the img element.