Use thesis/gcp-storage-bucket-action from alpine-version-413.0.0 branch
We temporarily change the version of `thesis/gcp-storage-bucket-action` code
used to publish the client image to the GCP. The version that we change to
builds the action using the `413.0.0-alpine` version of the `cloud-sdk`. We want
to do it because the latest version of the `cloud-sdk:alpine` resolves currently
to `414.0.0-alpine`, which intoduced a problem with execution of the `rsync`
command for requesters who do not have the `storage.buckets.get` permission. As
we use `rsync` in the `gcp-storage-bucket-action` action, we would be affected
by the issue if we would use the `v3.1.0` of the action (which uses latest
`cloud-sdk:alpine`).
More info:
https://cloud.google.com/storage/docs/release-notes#January_19_2023
GoogleCloudPlatform/gsutil#1663
thesis/gcp-storage-bucket-action@ee8b6a9