Change some fields to `markdown` widget
Some components such as buttons or nav links force the uppercase but in
some cases, we want to control it. So here we convert the button and nav
items to a `markdown` widget in cms config so we can force the lowercase
eg. to display `tBTC` we need to update the field in the admin panel as
follows: `<span style="text-transform: lowercase">t</span>BTC`.